fix: converge AccessDenied when fetching object to 404 NoSuchKey
The StaticBucketPolicy added to the s3 bucket, as part of Arc 10,
in lieu of ACLs, made it such that a 403 could be received from
the s3.getObject action:

> If you have the s3:ListBucket permission on the bucket, Amazon S3
will return an HTTP status code 404 ("no such key") error.
If you don’t have the s3:ListBucket permission,
Amazon S3 will return an HTTP status code 403 ("access denied") error.

So asap converges the 403 case, when fetching objects, to the 404 case